feat: move 'FakeBcryptAdapter' to adapters folder
